Move calculating the trend and selection into edit scan config dialog
Trend and selection are derived from already used data. Therefore they
should be calculated inside the dialog instead of passing them into it.
